+++ library/graphs/tree_path_queries.cpp
@@ -38,7 +38,7 @@ template <class T> struct LCAJumpSparse {
     }
 
     void dfs(int src = 0) {
-        for (int i = 1; i < par.size(); i++) {
+        for (int i = 1; i < (int)par.size(); i++) {
             par[i][src] = par[i - 1][par[i - 1][src]];
             sparse[i][src] = comb(sparse[i - 1][src], sparse[i - 1][par[i - 1][src]]);
         }
@@ -50,7 +50,7 @@ template <class T> struct LCAJumpSparse {
     }
 
     int jump(int x, int d) {
-        for (int i = 0; i < par.size(); i++) {
+        for (int i = 0; i < (int)par.size(); i++) {
             if ((d >> i) & 1) {
                 x = par[i][x];
             }
@@ -62,7 +62,7 @@ template <class T> struct LCAJumpSparse {
         if (depth[x] < depth[y]) std::swap(x, y);
         x = jump(x, depth[x] - depth[y]);
         if (x == y) return x;
-        for (int i = par.size() - 1; i >= 0; i--) {
+        for (int i = (int)par.size() - 1; i >= 0; i--) {
             int nx = par[i][x];
             int ny = par[i][y];
             if (nx != ny) x = nx, y = ny;
